Split read and write scenarios into separate examples
